requirements.txt
torch
torchvision
flask
Pillow
numpy
matplotlib
scikit-learn
